george/requirements.txt
psycopg2-binary
lxml
beautifulsoup4
tqdm
python-dotenv
openai
requests
tqdm
tiktoken
